+++ trview.app/UI/ViewOptions.h
@@ -26,7 +26,7 @@ namespace trview
         explicit ViewOptions(ui::Control& parent, const ITextureStorage& texture_storage);
 
         /// Event raised when an alternate group is toggled. The group number and the new state are passed as parameters.
-        Event<uint16_t, bool> on_alternate_group;
+        Event<uint32_t, bool> on_alternate_group;
 
         /// Event raised when the user has changed the depth of neighbour to display. The newly selected depth is passed
         /// when the event is raised.
@@ -130,7 +130,7 @@ namespace trview
         ui::Window* _tr1_3_panel;
         ui::Window* _tr4_5_panel;
         ui::StackPanel* _alternate_groups;
-        std::unordered_map<uint16_t, bool> _alternate_group_values;
-        std::unordered_map<uint16_t, ui::Button*> _alternate_group_buttons;
+        std::unordered_map<uint32_t, bool> _alternate_group_values;
+        std::unordered_map<uint32_t, ui::Button*> _alternate_group_buttons;
     };
 }
